Kill Duke and replace with Chagee
