+++ src/main/java/chagee/exception/EmptyDescriptionException.java
@@ -1,0 +1,10 @@
+package chagee.exception;
+
+/**
+ * A class to represent an empty description instantiation.
+ */
+public class EmptyDescriptionException extends Error {
+    public EmptyDescriptionException() {
+        super("☹ OOPS!!! The description of a task cannot be empty :-(");
+    }
+}
