+++ src/main/java/chagee/exception/InvalidEventException.java
@@ -1,0 +1,10 @@
+package chagee.exception;
+
+/**
+ * A class to represent an invalid Event instatiation.
+ */
+public class InvalidEventException extends Error {
+    public InvalidEventException() {
+        super("☹ OOPS!!! from date should be before to date!");
+    }
+}
