--- src/main/java/duke/command/DukeCommand.java
@@ -1,16 +1,0 @@
-package duke.command;
-
-/**
- * Enum wrapper for commands used by Duke.
- */
-public enum DukeCommand {
-    BYE("bye"), MARK("mark"), UNMARK("unmark"), DELETE("delete"), LIST("list"), TODO(
-            "todo"), DEADLINE(
-                    "deadline"), EVENT("event"), FIND("find"), VIEW_SCHEDULE("view_schedule");
-
-    public final String text;
-
-    private DukeCommand(String commandText) {
-        this.text = commandText;
-    }
-}
