--- text-ui-test/EXPECTED.TXT
@@ -1,68 +1,0 @@
-    ____________________________________________________________
-     Hello! I'm Duke
-     What can I do for you?
-    ____________________________________________________________
-    ____________________________________________________________
-    ____________________________________________________________
-    ____________________________________________________________
-     ☹ OOPS!!! The description of a todo cannot be empty.
-    ____________________________________________________________
-    ____________________________________________________________
-     Got it. I've added this task:
-       [T][ ] todo1
-
-     Now you have 1 tasks in the list.
-    ____________________________________________________________
-    ____________________________________________________________
-     Got it. I've added this task:
-       [T][ ] todo 2
-
-     Now you have 2 tasks in the list.
-    ____________________________________________________________
-    ____________________________________________________________
-     Got it. I've added this task:
-       [D][ ] homework
- (by: Oct 10 2023)
-     Now you have 3 tasks in the list.
-    ____________________________________________________________
-    ____________________________________________________________
-     ☹ OOPS!!! The time format is invalid, please use yyyy-MM-dd
-    ____________________________________________________________
-    ____________________________________________________________
-     ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
-    ____________________________________________________________
-    ____________________________________________________________
-     1.   [T][ ] todo1
- 
-     2.   [T][ ] todo 2
- 
-     3.   [D][ ] homework
- (by: Oct 10 2023) 
-    ____________________________________________________________
-    ____________________________________________________________
-     ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
-    ____________________________________________________________
-    ____________________________________________________________
-     ☹ OOPS!!! The time format is invalid, please use yyyy-MM-dd
-    ____________________________________________________________
-    ____________________________________________________________
-     ☹ OOPS!!! The time format is invalid, please use yyyy-MM-dd
-    ____________________________________________________________
-    ____________________________________________________________
-     ☹ OOPS!!! from date should be before to date!
-    ____________________________________________________________
-    ____________________________________________________________
-     Got it. I've added this task:
-       [E][ ] event two
- (from: Jan 31 2023 to: Jan 31 2023)
-     Now you have 4 tasks in the list.
-    ____________________________________________________________
-    ____________________________________________________________
-     Got it. I've added this task:
-       [E][ ] event two
- (from: Jan 30 2023 to: Jan 31 2023)
-     Now you have 5 tasks in the list.
-    ____________________________________________________________
-    ____________________________________________________________
-     Bye. Hope to see you again soon!
-    ____________________________________________________________
